separate out TelemetryCollection
